requirements.txt
kubernetes
boto3
pynacl
